Replaced most QVectors with std::vector

+++ engine/opengl/opengl_render_objects.cpp
@@ -133,7 +133,7 @@ void DrOpenGL::drawObject(DrEngineThing *thing, DrThingType &last_thing, bool dr
     DrEngineTexture *texture;
     long texture_number = object->getTextureNumber();
     if (object->animation_idle_keys.size() > 0) {
-        texture_number = object->animation_idle_keys[static_cast<size_t>(object->animation_idle_frame - 1)];
+        texture_number = object->animation_idle_keys[object->animation_idle_frame - 1];
     }
     texture = m_engine->getTexture(texture_number);
     if (texture == nullptr) return;
@@ -368,7 +368,7 @@ void DrOpenGL::drawObjectSimple(DrEngineThing *thing) {
     DrEngineTexture *texture;
     long texture_number = object->getTextureNumber();
     if (object->animation_idle_keys.size() > 0) {
-        texture_number = object->animation_idle_keys[static_cast<size_t>(object->animation_idle_frame - 1)];
+        texture_number = object->animation_idle_keys[object->animation_idle_frame - 1];
     }
     texture = m_engine->getTexture(texture_number);
     if (texture == nullptr) return;
@@ -467,7 +467,7 @@ bool DrOpenGL::drawObjectOccluder(DrEngineThing *thing, bool need_init_shader) {
     DrEngineTexture *texture;
     long texture_number = object->getTextureNumber();
     if (object->animation_idle_keys.size() > 0) {
-        texture_number = object->animation_idle_keys[static_cast<size_t>(object->animation_idle_frame - 1)];
+        texture_number = object->animation_idle_keys[object->animation_idle_frame - 1];
     }
     texture = m_engine->getTexture(texture_number);
     if (texture == nullptr) return true;
